Move TreeViewItemModel type to src/components/renders/tree-view/types.ts
